+++ autoload.php
@@ -13,7 +13,11 @@ if (!function_exists('di52_findFile')) {
 if (!function_exists('di52_autoload')) {
 	function di52_autoload($class)
 	{
-		include di52_findFile($class);
+		$file = di52_findFile($class);
+		if ($file === false) {
+			return false;
+		}
+		include $file;
 	}
 }